requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic[email]==2.9.2
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
httpx==0.27.2
pytest==8.3.3
pytest-cov==5.0.0
